Moved a LOT of other functionality into westtools/wipi.py
