+++ lib/west_tools/westtools/__init__.py
@@ -6,4 +6,4 @@ from selected_segs import SegSelector
 from binning import BinMappingComponent, mapper_from_dict
 from progress import ProgressIndicatorComponent
 from plot import Plotter
-from dtypes import WIPIDataset
+from wipi import WIPIDataset, KineticsIteration, __get_data_for_iteration__, WIPIScheme
